migrates to v2 format
