+++ examples/2_Accelerometer/2_Accelerometer.cpp
@@ -1,13 +1,16 @@
 /* -----------------------------------------------------------
-This example shows a lot of different features. As configured here
-it'll check for a good GPS fix every 10 minutes and publish that data
-if there is one. If not, it'll save you data by staying quiet. It also
-registers 3 Particle.functions for changing whether it publishes,
-reading the battery level, and manually requesting a GPS reading.
+This example shows some clever ways to use the accelerometer
+and report that data efficiently. It also includes the transmit
+mode and battery reading functions from the previous example.
+
+By default, it publishes only when there's a significant bump
+against it AND there's been at least 10 minutes since the last
+publish. NOTE: the first publish will only happen after the
+device has been on for at least this long!
 ---------------------------------------------------------------*/
 
 // Getting the library
-#include "AssetTracker/AssetTracker.h"
+#include "AssetTracker.h"
 
 // Set whether you want the device to publish data to the internet by default here.
 // 1 will Particle.publish AND Serial.print, 0 will just Serial.print
@@ -18,8 +21,12 @@ int transmittingData = 1;
 // Used to keep track of the last time we published data
 long lastPublish = 0;
 
-// How many minutes between publishes? 10+ recommended for long-time continuous publishing!
+// How many minutes minimum between publishes? 10+ recommended!
 int delayMinutes = 10;
+
+// Threshold to trigger a publish
+// 9000 is VERY sensitive, 12000 will still detect small bumps
+int accelThreshold = 12000;
 
 // Creating an AssetTracker named 't' for us to reference
 AssetTracker t = AssetTracker();
@@ -33,48 +40,39 @@ void setup() {
     // Sets up all the necessary AssetTracker bits
     t.begin();
     
-    // Enable the GPS module. Defaults to off to save power. 
-    // Takes 1.5s or so because of delays.
-    t.gpsOn();
-    
     // Opens up a Serial port so you can listen over USB
     Serial.begin(9600);
     
     // These three functions are useful for remote diagnostics. Read more below.
+    Particle.function("aThresh",accelThresholder);
     Particle.function("tmode", transmitMode);
     Particle.function("batt", batteryStatus);
-    Particle.function("gps", gpsPublish);
 }
 
 // loop() runs continuously
 void loop() {
-    // You'll need to run this every loop to capture the GPS output
-    t.updateGPS();
-
-    // if the current time - the last time we published is greater than your set delay...
-    if(millis()-lastPublish > delayMinutes*60*1000){
-        // Remember when we published
-        lastPublish = millis();
+    // Check if there's been a big acceleration
+    if(t.readXYZmagnitude() > accelThreshold ){
+        // Create a nice string with commas between x,y,z
+        String pubAccel = String::format("%d,%d,%d",t.readX(),t.readY(),t.readZ());
         
-        //String pubAccel = String::format("%d,%d,%d",t.readX(),t.readY(),t.readZ());
+        // Send that acceleration to the serial port where it can be read by USB
         //Serial.println(pubAccel);
-        //Particle.publish("A", pubAccel, 60, PRIVATE);
+        Serial.println(t.readXYZmagnitude());
         
-        // Dumps the full NMEA sentence to serial in case you're curious
-        Serial.println(t.preNMEA());
-        
-        // GPS requires a "fix" on the satellites to give good data,
-        // so we should only publish data if there's a fix
-        if(t.gpsFix()){
-            // Only publish if we're in transmittingData mode 1;
-            if(transmittingData){
-                // Short publish names save data!
-                Particle.publish("G", t.readLatLon(), 60, PRIVATE);
-            }
-            // but always report the data over serial for local development
-            Serial.println(t.readLatLon());
+        // If it's set to transmit AND it's been at least delayMinutes since the last one...
+        if(transmittingData && ((millis()-lastPublish) > (delayMinutes*60*1000))){
+            lastPublish = millis();
+            Particle.publish("A", pubAccel, 60, PRIVATE);
         }
+        
     }
+}
+
+// Remotely change the trigger threshold!
+int accelThresholder(String command){
+    accelThreshold = atoi(command);
+    return 1;
 }
 
 // Allows you to remotely change whether a device is publishing to the cloud
@@ -83,19 +81,6 @@ void loop() {
 int transmitMode(String command){
     transmittingData = atoi(command);
     return 1;
-}
-
-// Actively ask for a GPS reading if you're impatient. Only publishes if there's
-// a GPS fix, otherwise returns '0'
-int gpsPublish(String command){
-    if(t.gpsFix()){ 
-        Particle.publish("G", t.readLatLon(), 60, PRIVATE);
-        
-        // uncomment next line if you want a manual publish to reset delay counter
-        // lastPublish = millis();
-        return 1;
-    }
-    else { return 0; }
 }
 
 // Lets you remotely check the battery status by calling the function "batt"
