+++ examples/3_WakeOnMove/3_WakeOnMove.cpp
@@ -1,7 +1,7 @@
 
 #include "Particle.h"
 
-#include "AssetTracker/AssetTracker.h"
+#include "AssetTracker.h"
 
 // Project Location:
