+++ library.properties
@@ -1,0 +1,5 @@
+name=AssetTracker
+version=0.0.5
+license=LGPL
+author=Richard Whitney <richard@particle.io>
+sentence=Functions to make the Asset Tracker for Electron easy to use!
