--- spark.json
@@ -1,7 +1,0 @@
-{
-    "name": "AssetTracker",
-    "version": "0.0.5",
-    "author": "Richard Whitney <richard@particle.io>",
-    "license": "LGPL",
-    "description": "Functions to make the Asset Tracker for Electron easy to use!"
-}
