+++ src/AssetTracker.cpp
