+++ src/AssetTracker/AssetTracker.h
@@ -1,0 +1,1 @@
+#include "../AssetTracker.h"
